removed unnecessary files and imports
