+++ chemprop/data/__init__.py
@@ -2,5 +2,3 @@ from .data import MoleculeDatapoint, MoleculeDataset
 from .scaffold import cluster_split
 from .scaler import StandardScaler
 from .similarity import morgan_similarity, scaffold_similarity
-from .unsupervised_cluster import generate_unsupervised_cluster_labels
-from .vocab import get_vocab, load_vocab
