+++ chemprop/features/__init__.py
@@ -1,6 +1,5 @@
 from .featurization import atom_features, bond_features, BatchMolGraph, get_atom_fdim, get_bond_fdim, mol2graph
 from .functional_groups import FunctionalGroupFeaturizer
-from .kernels import get_kernel_func
 from .morgan_fingerprint import morgan_fingerprint
 from .rdkit_features import rdkit_2d_features
 from .utils import load_features, get_features_func
