+++ chemprop/models/model.py
@@ -5,11 +5,7 @@ from typing import Dict
 import torch.nn as nn
 import torch.nn.functional as F
 
-from .gan import GAN
-from .jtnn import JTNN
-from .moe import MOE
 from .mpn import MPN
-from .learned_kernel import LearnedKernel
 from chemprop.nn_utils import get_activation_function, initialize_weights, MayrDropout, MayrLinear, MAMLLinear
