+++ scripts/avg_dups.py
@@ -1,8 +1,56 @@
 from argparse import ArgumentParser
 import sys
 sys.path.append('../')
+from collections import defaultdict
+import numpy as np
 
-from chemprop.data_processing import average_duplicates
+from chemprop.data.utils import get_data, get_header
+
+def average_duplicates(args):
+    """Averages duplicate data points in a dataset."""
+    print('Loading data')
+    header = get_header(args.data_path)
+    data = get_data(path=args.data_path)
+    print(f'Data size = {len(data):,}')
+
+    # Map SMILES string to lists of targets
+    smiles_in_order = []
+    smiles_to_targets = defaultdict(list)
+    for smiles, targets in zip(data.smiles(), data.targets()):
+        smiles_to_targets[smiles].append(targets)
+        if len(smiles_to_targets[smiles]) == 1:
+            smiles_in_order.append(smiles)
+
+    # Find duplicates
+    duplicate_count = 0
+    stds = []
+    new_data = []
+    for smiles in smiles_in_order:
+        all_targets = smiles_to_targets[smiles]
+        duplicate_count += len(all_targets) - 1
+        num_tasks = len(all_targets[0])
+
+        targets_by_task = [[] for _ in range(num_tasks)]
+        for task in range(num_tasks):
+            for targets in all_targets:
+                if targets[task] is not None:
+                    targets_by_task[task].append(targets[task])
+
+        stds.append([np.std(task_targets) if len(task_targets) > 0 else 0.0 for task_targets in targets_by_task])
+        means = [np.mean(task_targets) if len(task_targets) > 0 else None for task_targets in targets_by_task]
+        new_data.append((smiles, means))
+
+    print(f'Number of duplicates = {duplicate_count:,}')
+    print(f'Duplicate standard deviation per task = {", ".join(f":{std:.4e}" for std in np.mean(stds, axis=0))}')
+    print(f'New data size = {len(new_data):,}')
+
+    # Save new data
+    with open(args.save_path, 'w') as f:
+        f.write(','.join(header) + '\n')
+
+        for smiles, avg_targets in new_data:
+            f.write(smiles + ',' + ','.join(str(value) if value is not None else '' for value in avg_targets) + '\n')
+
 
 if __name__ == '__main__':
     parser = ArgumentParser()
